+++ .eslintrc.js
@@ -48,5 +48,6 @@ module.exports = {
     '@sentry-internal/sdk/no-optional-chaining': 'off',
     '@sentry-internal/sdk/no-nullish-coalescing': 'off',
     '@sentry-internal/sdk/no-class-field-initializers': 'off',
+    'deprecation/deprecation': 'off',
   },
 };