+++ CHANGELOG.md
@@ -8,9 +8,9 @@
 
 ### Dependencies
 
-- Bump Cocoa SDK from v8.18.0 to v8.22.0 ([#555](https://github.com/getsentry/sentry-capacitor/pull/555))
-  - [changelog](https://github.com/getsentry/sentry-cocoa/blob/main/CHANGELOG.md#8220)
-  - [diff](https://github.com/getsentry/sentry-cocoa/compare/8.18.0...8.22.0)
+- Bump Cocoa SDK from v8.18.0 to v8.22.2 ([#555](https://github.com/getsentry/sentry-capacitor/pull/555), [#604](https://github.com/getsentry/sentry-capacitor/pull/604))
+  - [changelog](https://github.com/getsentry/sentry-cocoa/blob/main/CHANGELOG.md#8222)
+  - [diff](https://github.com/getsentry/sentry-cocoa/compare/8.18.0...8.22.2)
 
 ## 0.16.0