+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-### Fixes/Features
+### Dependencies
 
 - Bump JavaScript SDK from v7.93.0 to v7.114.0 ([#645](https://github.com/getsentry/sentry-capacitor/pull/645))
   - [changelog](https://github.com/getsentry/sentry-javascript/blob/7.114.0/CHANGELOG.md)